Backend err handle and delete post + comment

+++ backend/api_layer/api.py
@@ -2,7 +2,7 @@ import logging
 import os
 import psycopg
 from psycopg.errors import ProgrammingError
-from flask import Flask, request, json
+from flask import Flask, request, json, abort
 from flask_ngrok import run_with_ngrok
 import json
 import atexit    # Gracefully exit and cleanup after normal termination
@@ -85,6 +85,9 @@ def insert_user():
     name = data.get('name')
     avatar = data.get('avatar')
 
+    if not name:
+        abort(400, "name field is missing")
+
     print(name)
 
     id = DBService.insert_user(connection, name, avatar)
@@ -104,7 +107,9 @@ def get_user():
         data = json.loads(request.data)
     
     id = data.get('id')
-    print(id)
+    
+    if not id:
+        abort(400, "id field is missing")
 
     user = DBService.getUser(connection, id)
     print(user)
@@ -133,6 +138,9 @@ def update_avatar():
     id = data.get('id')
     avatar = data.get('avatar')
 
+    if not (id and avatar):
+        abort(400, "missing required field")
+
     result = DBService.updateAvatar(connection, id, avatar)
 
     return {"updated": result} 
@@ -151,6 +159,9 @@ def update_name():
     id = data.get('id')
     name = data.get('name')
 
+    if not (id and name):
+        abort(400, "missing required field")
+
     result = DBService.updateName(connection, id, name)
 
     return {"updated": result} 
@@ -167,6 +178,9 @@ def addToFavorite():
 
     uid = data.get("id")
     mid = data.get("movie_id")
+
+    if not (uid and mid):
+        abort(400, "missing required field")
 
     result, msg = DBService.addToFavoriteMovie(connection, uid, mid)
 
@@ -185,6 +199,9 @@ def removeFromFavorite():
     uid = data.get("id")
     mid = data.get("movie_id")
 
+    if not (uid and mid):
+        abort(400, "missing required field")
+
     result = DBService.removeFromFavoriteMovie(connection, uid, mid)
 
     return {"removed:": result}
@@ -202,9 +219,12 @@ def addFriend():
     uid1 = data.get("uid1")
     uid2 = data.get("uid2")
 
+    if not (uid1 and uid2):
+        abort(400, "missing required field")
+
     result, msg = DBService.addFriend(connection, uid1, uid2)
 
-    return {"added:": result, "message": msg}
+    return {"added": result, "message": msg}
 
 @app.route('/user/Friend', methods=['DELETE'])
 def removeFriend():
@@ -218,6 +238,9 @@ def removeFriend():
 
     uid1 = data.get("uid1")
     uid2 = data.get("uid2")
+
+    if not (uid1 and uid2):
+        abort(400, "missing required field")
 
     result = DBService.removeFriend(connection, uid1, uid2)
 
@@ -239,17 +262,18 @@ def createWatchParty():
     time = data.get("time")
     platform = data.get("platform")
 
-    if mid == None:
-        return {"created:": False, "message": "Creating a new WatchParty with no movie is not allowed."}
-    if date == None or time == None:
-        return {"created:": False, "message": "Creating a new WatchParty with insufficient time information is not allowed."}
+    if not (uid and mid and date and time and platform):
+        abort(400, "missing required field")
     
     if not is_valid_date(date):
-        return {"created:": False, "message": "Invalid date format"}
+        abort(400, "Invalid date format")
 
     result, msg, wid = DBService.createWatchParty(connection, uid, mid, date, time, platform)
 
-    return {"created:": result, "message" : msg, "wid": wid}
+    if not result:
+        abort(400, msg)
+
+    return {"created": result, "wid": wid}
 
 @app.route('/user/WatchParty', methods=['GET'])
 def getWatchParty():
@@ -261,12 +285,15 @@ def getWatchParty():
         print("received non-json object, converting to json")
         data = json.loads(request.data)
 
-    organizer = data.get("uid")
+    organizer = data.get("id")
+
+    if not organizer:
+        abort(400, "required field empty")
 
     queryresult = DBService.getUserWatchParty(connection, organizer)
 
     if queryresult == None:
-        return {"message:": "No watch party is created by this user"}
+        return {"message": "No watch party is created by this user"}
     
     result = []
 
@@ -292,6 +319,9 @@ def deleteWatchParty():
 
     wid = data.get("wid")
 
+    if not wid:
+        abort(400, "missing required field")
+
     result = DBService.deleteWatchParty(connection, wid)
 
     return {"removed": result}
@@ -306,12 +336,15 @@ def participateWatchParty():
         print("received non-json object, converting to json")
         data = json.loads(request.data)
 
-    participant = data.get("participantID")
+    participant = data.get("uid")
     wid = data.get("wid")
 
-    result = DBService.addToWatchParty(connection, participant, wid)
+    if not (participant and wid):
+        abort(400, "missing required field")
 
-    return {"added:": result}
+    result, msg = DBService.addToWatchParty(connection, participant, wid)
+
+    return {"added": result, "message": msg}
 
 @app.route('/user/joinWatchPartyWithNoID', methods=['POST'])
 def participateWatchPartyWithNoId():
@@ -323,16 +356,19 @@ def participateWatchPartyWithNoId():
         print("received non-json object, converting to json")
         data = json.loads(request.data)
 
-    participant = data.get("participantID")
+    participant = data.get("participantid")
     organizer = data.get("organizerID")
     movie = data.get("mid")
     date = data.get("date")
+
+    if not (participant and organizer and movie and date):
+        abort(400, "missing required field")
 
     result = DBService.addToWatchPartyWithNoId(connection, participant, organizer, movie, date)
 
     # TODO: make (organizer, mid, date) unique in sql table: UNIQUE(col1, col2, col3)
 
-    return {"added:": result}
+    return {"added": result}
 
 @app.route('/user/getParticipatedWatchParty', methods=['GET'])
 def getParticipatedWatchParty():
@@ -344,21 +380,29 @@ def getParticipatedWatchParty():
         print("received non-json object, converting to json")
         data = json.loads(request.data)
 
-    participant = data.get("participantID")
+    participant = data.get("uid")
+
+    if not participant:
+        abort(400, "missing required field")
 
     queryresult = DBService.getUserParticipatedWatchParty(connection, participant)
 
     if queryresult == None:
-        return {"message:": "This user have no participated Watch Party"}
+        return {"message": "This user have no participated Watch Party"}
     
     result = []
+
+    print(queryresult)
 
     for watchParty in queryresult:
         result.append({"wid": watchParty[0],
                 "ownerId": watchParty[1],
                 "movieId": watchParty[2],
                 "Date": watchParty[3],
-                "Platform": watchParty[4]
+                "time": watchParty[4].strftime('%H:%M'),
+                "Platform": watchParty[5],
+                "owner's Name": watchParty[6],
+                "owner's avatar": watchParty[7]
                 })
     
     return {"result": result}
@@ -378,6 +422,9 @@ def createPost():
     movie_id = data.get("movie_id", 0)
     content = data.get("content")
 
+    if not (uid and movie_id and content):
+        abort(400, "missing required field")
+
     # Optional
     newWatchParty = data.get("newWatchParty", False)   # optional. true -> create new watchparty
     wid = 0
@@ -387,13 +434,19 @@ def createPost():
     if (newWatchParty):
         date = data.get("date")
         platform = data.get("platform")
+        time = data.get("time")
 
-        if movie_id == None:
-            return {"message:": "Creating a new WatchParty with no movie is not allowed."}
+        if not (date and platform and time):
+            abort(400, "missing required field")
 
         # TODO: check valid date string type (refer to CockroachDB). Return false if invalid
+        if not is_valid_date(date):
+            abort(400, "Invalid date format")
 
-        wid = DBService.createWatchParty(connection, uid, movie_id, date, platform)
+        ok, msg, wid = DBService.createWatchParty(connection, uid, movie_id, date, time, platform)
+
+        if not ok:
+            abort(400, msg)
     
     pid = DBService.createPost(connection, uid, movie_id, content, wid)
 
@@ -412,17 +465,25 @@ def getUserPost():
     # Required Fields:
     uid = data.get("uid")
 
+    if not uid:
+        abort(400, "missing required field")
+
     posts = DBService.getUserPost(connection, uid)
 
     result = []
+
+    print("post:", posts)
 
     for p in posts:
         result.append({"pid": p[0],
                        "writer": p[1],
                        "movie_id": p[2],
                        "pdate": p[3],
-                       "content": p[4],
-                       "wid": p[5]
+                       "ptime": p[4].strftime('%H:%M'),
+                       "content": p[5],
+                       "wid": p[6],
+                       "writer's name": p[7],
+                       "writer's avatar": p[8]
             })
         
     return {"result": result}
@@ -441,6 +502,9 @@ def getAllPost():
     # Required Fields:
     uid = data.get("uid")
 
+    if not uid:
+        abort(400, "missing required field")
+
     posts = DBService.getAllPost(connection, uid)
 
     result = []
@@ -450,11 +514,34 @@ def getAllPost():
                        "writer": p[1],
                        "movie_id": p[2],
                        "pdate": p[3],
-                       "content": p[4],
-                       "wid": p[5]
+                       "ptime": p[4].strftime('%H:%M'),
+                       "content": p[5],
+                       "wid": p[6],
+                       "writer's name": p[7],
+                       "writer's avatar": p[8]
             })
         
     return {"result": result}
+
+@app.route('/post', methods=['DELETE'])
+def deletePost():
+    data = {}
+
+    if request.is_json:
+        data = request.json
+    else:
+        print("received non-json object, converting to json")
+        data = json.loads(request.data)
+
+    # Required Fields:
+    pid = data.get("pid")
+
+    if not pid:
+        abort(400, "missing required field")
+
+    ok = DBService.deletePost(connection, pid)
+
+    return {"deleted": ok}
 
 
 @app.route("/comment", methods=["POST"])
@@ -472,6 +559,9 @@ def postComment():
     pid = data.get("pid")
     content = data.get("content")
 
+    if not (uid and pid and content):
+        abort(400, "missing required field")
+
     cid = DBService.postComment(connection, uid, pid, content)
 
     return {"posted": True, "cid": cid}
@@ -488,6 +578,8 @@ def getComment():
 
     # Required Fields:
     pid = data.get("pid")
+    if not pid:
+        abort(400, "missing required field")
 
     comments = DBService.getComment(connection, pid)
 
@@ -496,14 +588,36 @@ def getComment():
     for c in comments:
         result.append({
             "cid": c[0],
-            "content": c[1],
-            "cdate": c[2],
-            "uid": c[3],
-            "pid": c[4]
+            "pid": c[1],
+            "content": c[2],
+            "date": c[3],
+            "time": c[4].strftime('%H:%M'),
+            "user": c[5],
+            "user name": c[6],
+            "user's avatar": c[7]
         })
 
     return {"result":result}
 
+
+@app.route("/comment", methods=["DELETE"])
+def deleteComment():
+    data = {}
+
+    if request.is_json:
+        data = request.json
+    else:
+        print("received non-json object, converting to json")
+        data = json.loads(request.data)
+
+    # Required Fields:
+    cid = data.get("cid")
+    if not cid:
+        abort(400, "missing required field")
+
+    ok = DBService.deleteComment(connection, cid)
+
+    return {"deleted": ok}
 
 if __name__ == '__main__':
     # app.run(debug=True)
